+++ src/core/recipe-book.ts
@@ -1,11 +1,7 @@
 import type { Ingredient } from "./ingredient";
+import type { Yield } from "./recipe-yield";
 
 /* ---- TYPES ---- */
-
-export type Yield = {
-  amount: number;
-  units: string;
-};
 
 export type Recipe = {
   yield: Yield;
